simulate_one: activity time breakdown

+++ src/timespan.cpp
@@ -2,7 +2,7 @@
 
 #include "timespan.hpp"
 
-timespan timespan_calculate(usize const secs)
+timespan timespan_calculate(u64 const secs)
 {
   usize const SECONDS_PER_DAY = 86400, SECONDS_PER_HOUR = 3600, SECONDS_PER_MINUTE = 60;
