+++ src/timespan.hpp
@@ -8,10 +8,10 @@
 #include "primitives.hpp"
 
 struct timespan {
-  usize days, hours, minutes, seconds;
+  u64 days, hours, minutes, seconds;
 };
 
-timespan timespan_calculate(usize secs);
+timespan timespan_calculate(u64 secs);
 std::string timespan_to_string(timespan const &ts);
 
 #endif // TIMESPAN_HPP
